+++ health-services/project/CHANGELOG.md
@@ -1,7 +1,8 @@
 All notable changes to this module will be documented in this file.
 
 ## 1.1.2 - 2024-02-26
-- Added Project start date and end date update validation
+- Implemented validation for updating project start date and end date.
+- Added numberOfSessions field in additional details for attendance registry.
 
 ## 1.1.1 - 2023-11-15
 - Added tag in project beneficiary